requirements/requirements-base.txt
Django>=2.1,<3.0
djangorestframework>=3.9.0,<4.0.0
django-encrypted-model-fields>=0.5.8,<1.0.0
django-guardian>=1.4.9,<2.0.0